# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Smart ERP API
  version: 1.1.0
  summary: Machine form of docs/spec/04-api-contracts.md
  description: |
    Seeded from `docs/spec/04-api-contracts.md` version 1.1.0. That file is the human form and
    wins on conflict until this document is regenerated. Additive changes bump the minor version;
    breaking changes bump the major version and carry a migration note.

    1.1.0 (additive): `GET /me/sessions` and `DELETE /me/sessions/{id}` plus schema `UserSession`
    (R1.15, A9). Period soft-close, hard-close, audit-adjustment open, and the holiday calendar.
    No breaking change from 1.0.0.

    Transport rules that apply to every operation:

    - JSON over HTTPS. All timestamps are UTC ISO 8601. Money is `{ amount, currency }` with the
      amount as a decimal string.
    - Every success response is `{ data, meta }`. Every failure, including 429 and gateway errors,
      is `{ error: { code, message, details, request_id } }`.
    - Every mutating request accepts `Idempotency-Key` (UUID); the server replays a stored response
      for 24 hours; a different body with the same key is `409 CONFLICT`.
    - Every mutable resource carries `state_version`; mutating requests send `If-Match`; a mismatch
      is `409 CONFLICT` with the current state in `error.details`.
    - `Accept-Language: en | ar` affects messages and print formats, never data.
  contact:
    name: Smart ERP contract owner
  license:
    name: Proprietary
...
servers:
...
security:
  - bearer: []
    dpop: []
...
paths:
...
  /approvals/inbox:
    get:
      operationId: listApprovalInbox
      tags: [approvals]
      summary: Approval inbox cards
      description: Cards for the authenticated approver, filtered by state (R2.10 fraud hints included).
      parameters:
        - name: state
          in: query
          required: false
          schema:
            type: string
            enum: [needs_me, waiting_on_others, fyi]
            default: needs_me
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/AcceptLanguage"
      responses:
        "200":
          description: Cards
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ApprovalCard"
                  meta:
                    $ref: "#/components/schemas/PageMeta"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/Unavailable"
  /approvals/{id}:
    get:
      operationId: getApproval
      tags: [approvals]
      summary: Live view of one approval request
      description: Includes the document snapshot at submission and its content hash (R2.11).
      parameters:
        - $ref: "#/components/parameters/ApprovalId"
        - $ref: "#/components/parameters/AcceptLanguage"
      responses:
        "200":
          description: Approval request
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    $ref: "#/components/schemas/ApprovalDetail"
                  meta:
                    $ref: "#/components/schemas/Meta"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/Unavailable"
  /approvals/{id}/approve:
    post:
      operationId: approveApproval
      tags: [approvals]
      summary: Approve
      description: |
        Serialised per request with a row lock and state version (R2.6). A concurrent second
        decision receives 200 with the decided state and `ALREADY_DECIDED` in `meta.notice`, not an
        error (D7). Step-up required at or above threshold (D14).
      parameters:
        - $ref: "#/components/parameters/ApprovalId"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/AcceptLanguage"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [state_version]
              properties:
                comment:
                  type: string
                  maxLength: 2000
                state_version:
                  $ref: "#/components/schemas/StateVersion"
                step_up_token:
                  type: string
      responses:
        "200":
          description: Decision recorded, or already decided
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    $ref: "#/components/schemas/ApprovalDecision"
                  meta:
                    $ref: "#/components/schemas/DecisionMeta"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/Unavailable"
  /approvals/{id}/reject:
    post:
      operationId: rejectApproval
      tags: [approvals]
      summary: Reject
      description: An empty or whitespace reason is `400 VALIDATION_ERROR` (R2.4, D5).
      parameters:
        - $ref: "#/components/parameters/ApprovalId"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/AcceptLanguage"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason, state_version]
              properties:
                reason:
                  type: string
                  minLength: 1
                  maxLength: 2000
                  pattern: "\\S"
                state_version:
                  $ref: "#/components/schemas/StateVersion"
      responses:
        "200":
          description: Decision recorded, or already decided
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    $ref: "#/components/schemas/ApprovalDecision"
                  meta:
                    $ref: "#/components/schemas/DecisionMeta"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/Unavailable"
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: ERP access token issued by POST /auth/token. Expires in 15 minutes or less.
    dpop:
      type: apiKey
      in: header
      name: DPoP
      description: |
        DPoP proof JWT signed by the enrolled device key (RFC 9449). Required on every mobile
        request alongside the bearer token; the console uses a session cookie with a CSRF token.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        UUID chosen by the client. The server stores the response for 24 hours and replays it for a
        repeated key with the same body; a different body with the same key is `409 CONFLICT`.
      schema:
        type: string
        format: uuid
    IfMatch:
      name: If-Match
      in: header
      required: false
      description: |
        The `state_version` the client last saw. A mismatch is `409 CONFLICT` with the current
        state in `error.details.current`.
      schema:
        type: string
    AcceptLanguage:
      name: Accept-Language
      in: header
      required: false
      description: "`en` or `ar`. Affects messages and print formats, never data."
      schema:
        type: string
        enum: [en, ar]
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque cursor from `meta.next_cursor` of the previous page.
      schema:
        type: string
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
...
    ApprovalId:
      name: id
      in: path
      required: true
      description: Approval request id
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Validation failed (`VALIDATION_ERROR`) or a required header is missing
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Unauthorized:
      description: "`AUTH_REQUIRED`, `TOKEN_EXPIRED`, `TOKEN_REVOKED` or `DEVICE_MISMATCH`"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Forbidden:
      description: "`PERMISSION_DENIED`, `STEP_UP_REQUIRED` or `SOD_VIOLATION`"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: "`NOT_FOUND`"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Conflict:
      description: |
        `CONFLICT` (stale `If-Match`, current state in `error.details.current`, or a reused
        `Idempotency-Key` with a different body), `IMMUTABLE` or `PERIOD_CLOSED`.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    RateLimited:
      description: "`RATE_LIMITED`, in the standard envelope (A7, A17)"
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds until the client may retry
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    InternalError:
      description: "`INTERNAL_ERROR`; the message never leaks internals, `request_id` correlates logs"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Unavailable:
      description: "`INTERNAL_ERROR` or `MISSING_CONFIG` while the service is not ready; also returned by the gateway in the same envelope"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    Money:
      type: object
      description: Decimal amount as a string plus an ISO 4217 currency code. Never a float.
      required: [amount, currency]
      properties:
        amount:
          type: string
          pattern: "^-?[0-9]+(\\.[0-9]+)?$"
          examples: ["1234.56"]
        currency:
          type: string
          pattern: "^[A-Z]{3}$"
          examples: ["AED"]
    Meta:
      type: object
      description: Present on every success response.
      required: [as_of, request_id]
      properties:
        as_of:
          type: string
          format: date-time
          description: Server time the response was computed, UTC
        request_id:
          type: string
          description: Correlates the response with logs and audit rows
      additionalProperties: true
    PageMeta:
      description: Meta for list responses. `total` is present only where cheap.
      allOf:
        - $ref: "#/components/schemas/Meta"
        - type: object
          properties:
            next_cursor:
              type: [string, "null"]
            total:
              type: integer
              minimum: 0
    DecisionMeta:
      description: Meta for approval decisions. `notice` carries `ALREADY_DECIDED` when a concurrent decision won (D7).
      allOf:
        - $ref: "#/components/schemas/Meta"
        - type: object
          properties:
            notice:
              type: string
              enum: [ALREADY_DECIDED]
    ErrorCode:
      type: string
      description: The closed set of error codes from docs/spec/04-api-contracts.md.
      enum:
        - AUTH_REQUIRED
        - TOKEN_EXPIRED
        - TOKEN_REVOKED
        - DEVICE_MISMATCH
        - STEP_UP_REQUIRED
        - PERMISSION_DENIED
        - VALIDATION_ERROR
        - MISSING_CONFIG
        - NOT_FOUND
        - CONFLICT
        - ALREADY_DECIDED
        - IMMUTABLE
        - PERIOD_CLOSED
        - SOD_VIOLATION
        - CREDIT_HOLD
        - PRICE_FLOOR_HOLD
        - NEGATIVE_STOCK
        - RATE_LIMITED
        - INTERNAL_ERROR
    Error:
      type: object
      required: [code, message, details, request_id]
      properties:
        code:
          $ref: "#/components/schemas/ErrorCode"
        message:
          type: string
          description: Human readable, localised by Accept-Language
        details:
          type: object
          description: Code-specific detail, e.g. `{ field }` for VALIDATION_ERROR or `{ current }` for CONFLICT
          additionalProperties: true
        request_id:
          type: string
    ErrorEnvelope:
      type: object
      description: Returned on every failure path including 429 and gateway errors (A17).
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/Error"
...
    StateVersion:
      type: integer
      minimum: 0
      description: Optimistic concurrency version; bumped on every transition
...
    Actor:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
        name:
          type: string
    ApprovalCard:
      type: object
      required:
        - request_id
        - doc_type
        - doc_number
        - amount
        - requester
        - waiting_since
        - severity
        - fraud_hints
        - allowed_actions
        - state_version
        - deep_link
      properties:
        request_id:
          type: string
        doc_type:
          type: string
        doc_number:
          type: string
        party:
          type: [string, "null"]
        amount:
          $ref: "#/components/schemas/Money"
        requester:
          $ref: "#/components/schemas/Actor"
        waiting_since:
          type: string
          format: date-time
        severity:
          $ref: "#/components/schemas/Severity"
        fraud_hints:
          type: array
          items:
            type: string
        allowed_actions:
          type: array
          items:
            $ref: "#/components/schemas/AllowedAction"
        state_version:
          $ref: "#/components/schemas/StateVersion"
        deep_link:
          type: string
          pattern: "^smarterp://"
    ApprovalDetail:
      allOf:
        - $ref: "#/components/schemas/ApprovalCard"
        - type: object
          required: [state, snapshot, snapshot_hash]
          properties:
            state:
              $ref: "#/components/schemas/ApprovalState"
            snapshot:
              type: object
              description: The document as submitted
              additionalProperties: true
            snapshot_hash:
              type: string
              description: sha256 of the canonical snapshot (R2.11); shown to approvers
              pattern: "^[a-f0-9]{64}$"
            decisions:
              type: array
              items:
                type: object
                required: [by, at, decision, hash_seen]
                properties:
                  by:
                    $ref: "#/components/schemas/Actor"
                  at:
                    type: string
                    format: date-time
                  decision:
                    type: string
                    enum: [approved, rejected, delegated]
                  comment:
                    type: string
                  hash_seen:
                    type: string
    ApprovalState:
      type: string
      enum: [pending, approved, rejected, delegated, snoozed, posted, cancelled]
    ApprovalDecision:
      type: object
      required: [request_id, state, state_version]
      properties:
        request_id:
          type: string
        state:
          $ref: "#/components/schemas/ApprovalState"
        state_version:
          $ref: "#/components/schemas/StateVersion"
        decided_by:
          $ref: "#/components/schemas/Actor"
        decided_at:
          type: string
          format: date-time
    Severity:
      type: string
      enum: [LOW, MEDIUM, HIGH, CRITICAL]
    AllowedAction:
      type: string
      enum: [approve, reject, acknowledge, snooze, delegate, open]
...